.history/MVP/MVP/requirements_20240626103814.txt
streamlit 
requests 
pandas 
matplotlib 
plotly 
google-generativeai


pdfminer
PyPDF2